JavaFX now be able to compile to jar executable jar file =)
