+++ dependency-reduced-pom.xml
@@ -139,6 +139,14 @@
       <scope>test</scope>
     </dependency>
   </dependencies>
+  <reporting>
+    <outputDirectory>C:/Users/apert/SomeWritableDirectory/target/site</outputDirectory>
+    <plugins>
+      <plugin>
+        <artifactId>maven-project-info-reports-plugin</artifactId>
+      </plugin>
+    </plugins>
+  </reporting>
   <properties>
     <maven.compiler.target>17</maven.compiler.target>
     <maven.compiler.source>17</maven.compiler.source>
