+++ src/main/java/umleditor/EntryPoint.java
@@ -1,27 +1,13 @@
 package umleditor;
 
 /*
- * This is the Main entry for both Sprint 1 and 2 comment out if you working with your correct packages
+ * This is the Main entry for both Sprint 1 and 2 comment out if your working with your correct packages
  */
 
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
 import umleditor.sprint1.utilities.Display;
+import umleditor.sprint2.view.AppStart;
 
-public class EntryPoint extends Application {
-
-  @Override
-  public void start(Stage primaryStage) throws Exception {
-    FXMLLoader loader = new FXMLLoader(getClass().getResource("/sprint2/CodeCain.fxml"));
-    Scene scene = new Scene(loader.load(), 900, 600);
-
-    primaryStage.setResizable(true);
-    primaryStage.setTitle("Code Cain Demo");
-    primaryStage.setScene(scene);
-    primaryStage.show();
-  }
+public class EntryPoint {
 
   public static void main(String[] args) {
 
@@ -34,6 +20,6 @@ public class EntryPoint extends Application {
     /*
     Sprint 2 Launch Codes
      */
-    launch(args);
+    AppStart.main(args);
   }
 }
