--- src/main/java/umleditor/sprint2/Controller.java
@@ -1,5 +1,0 @@
-package umleditor.sprint2;
-
-public class Controller {
-
-}
