+++ src/main/java/umleditor/sprint2/controller/Controller.java
@@ -1,0 +1,5 @@
+package umleditor.sprint2.controller;
+
+public class Controller {
+
+}
