+++ src/main/java/umleditor/sprint2/demo/CodeCainController.java
@@ -1,8 +1,14 @@
 package umleditor.sprint2.demo;
 
+
+import javafx.application.Application;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
 import javafx.scene.control.TextField;
+import javafx.stage.Stage;
+
 
 public class CodeCainController {
