+++ src/main/java/umleditor/sprint2/model/Model.java
@@ -1,0 +1,5 @@
+package umleditor.sprint2.model;
+
+public class Model {
+
+}
