+++ src/main/java/umleditor/sprint2/view/View.java
@@ -1,0 +1,5 @@
+package umleditor.sprint2.view;
+
+public class View {
+
+}
